Changed _StatisticalInfoTable definition so as to exploit move semantics
_StatisticalInfoTable was changed from vector<_StatisticalInfo> to vector<unique_ptr<_StatisticalInfo>>.
This is particularly useful as the results from multiple StatisticsGatherer are collected and returned
inside CompositeStatisticsGatherer objects. All the statistical info vectors in each StatisticsGatherer
need not to be copied inside contiguous area of memory, instead their ownership could be moved to the
composite results vector.
Hopefully, I get rid of all (or most of) the expensive copies.