Optimized SDEs calculation by fixing a default minimum time step.
A minimum time step of one day is established. This is the default time-delta to be used in
SDEs update calculation: another advantage is that more quantities in the SDEs can be precomputed.
The test payoff has been changed so as to reduce to a 1Y vanilla CALL, allowing for consistency checks
against the results obtained with Black-Scholes analytical solution.