Fixed imprecision in path range determination
The code for PathDependentScenarioSimulator has been updated to handle the case
of observers monitoring the last date in the generated path. A termination condition was added
to prevent invoking uselessly a path element update if the final date had already been reached.
This fixes finally guarantees full consistency in the price of a European Call option as calculated
using the path observers approach with a limit-case asian payoff and then the plain state payoff
of European options!