Exploited template specialization in PlainVanillaPayoff class, updated AsianPayoff accordingly
Template specialization of PlainVanillaPayoff based on the option side is more efficient than
runtime lookup at the option side: the construction of the state payoff to pass to MonteCarloEngine
gets, alas, more involved as the PlainVanillaPayoff objects should be known at compile time.
AsianPayoff was adapted to not depend on the new implementation of PlainVanillaPayoff.
Moreover, the final payoff evaluation has been changed so as to avoid a cycle of construction
and destruction of PlainVanillaPayoff objects.