Improved Timer utility class
Exploited template specification to get a flexible timer class allowing to measure seconds,
milliseconds and nanoseconds: if required, durations in arbitrary format can be measured with
slight modification of the class. Enabled Start and Stop method to the stopwatch.
Defined macro to instantiate a timer object associated to a specified variable name.