Started massive refactoring of the simulation framework to price path dependent payoffs
This is just the first step towards a new framework to handle the Monte Carlo simulations for
path dependent derivatives. The observer pattern framework has been replaced by a simpler one,
in which the relevant spot prices are passed from the MC engine to the payoff evaluator.
Many thigs in this implementation are just temporary crafts to make everything work.