Replaced pointer to payoff function to std::function wrapper in MonteCarloEngine class
It has been exploited the fact that only the payoff function is required by the MonteCarlo engine.
In this way we get rid of any complication due to polymorphic types for path dependent payoffs,
and we can copy the function wrappers by value preparing the ground for a multithreaded implementation
of the MonteCarlo routine.