Added an explicit class GaussianVariatesGenerator
The need to set a different rng seed for each execution thread led to move the normal variate generator
function in a specific class (as the FunctionalWrapperGenerator does not allow to update the generator seed).
This is probably a temporary fix, waiting for a more structured and thorough handing of the random variate generators.