Sketched a first multi-thread implementation of the MonteCarlo simulation
This version seems to work, however it has lots of drawbacks. First of all, different results are produced at
each run of the program: could this be due to the setting of the seed in each thread?
Moreover, it's ugly how the current implementation gathers all the results from each thread into a syntethic value.