Removed distinction between path-dependent and "State" payoffs
State payoffs (i.e. depending only on the spot value at maturity) are now treated as special case of generic
path dependent payoffs. This would probably do not affect performance that much (a vector now is used instead of
a simple double in the payoff function) but it drastically simplifies code.