Added evaluation of delta and gamma greeks, and performed other code improvements
Massive update. Delta and gamma are evaluated in the MC routine together with the price.
IMPORTANT:
A new approach is used to retrieve the pricing results, which is far less abstract and modifyable
than the previous and more concrete, dirt and implementation-specific.
This could raise problem if (e.g. in debug mode) we would like to gather more statistics on the results,
and can also pose problems if other greeks should be computed; especially if we want to enable the evaluation
on demand, i.e. calculating the greeks only if requested.