Turned EQPricer into a shared library
Important milestone: the project is converted into a shared library, which can be called by external codes
through APIs compliants with the C calling conventions. Only the API symbols are exported in the dll, without
name mangling. The export macro for the functions to expose is automatically generated by Cmake.
The external calling code would be e.g. a GUI or a script handling the pricing inputs and the results from the library.